+++ esp32p4-hal/examples/hello_world.rs
@@ -2,7 +2,10 @@
 #![no_main]
 
 use esp32p4_hal::{
-    clock::ClockControl, peripherals::Peripherals, prelude::*, system::SystemExt, Delay,
+    clock::{ClockControl, CpuClock},
+    peripherals::Peripherals,
+    prelude::*,
+    system::SystemExt,
 };
 use esp_backtrace as _;
 
@@ -10,7 +13,7 @@ use esp_backtrace as _;
 fn main() -> ! {
     let peripherals = Peripherals::take();
     let system = peripherals.SYSTEM.split();
-    let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock360MHz).freeze();
+    let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock90MHz).freeze();
 
     esp_println::println!("Hello, world!");
     loop {}